// Nominas/ClientApp/src/app/nomina/models/persona.ts
export class Persona {
  identificacion: number;
  nombres: string;
  apellidos: string;
  sexo: string;
  fechadenacimiento: Date;
  edad: number;
  fechaingreso: Date;
  tipoContrato: string;
  estado: string;
  fechaNacimiento: string | number | Date;
  diaslab: number;
  vrNeto: number;
  vrDia: number;
  Firma: string;
  pago: number;

}